Initialized AssetManger development
AssetManager class is to be used to manage all assets for the app.
Currently the implementation is done to transfer the tessdata folder
for the TessBaseApi to the internal storage for the app. This creates a
redundant ~37MB in internal disk storage usage for the english language
data but necessary for the ocr api, since we will store the english
data in the app apk to make sure we always have it.